create interface and split anim data from interface
