+++ ReAnimation_v2.omwscripts
@@ -1,1 +1,2 @@
-PLAYER: scripts\MaxYari\ReAnimation_v2\ReAnimation_v2.lua
+PLAYER NPC: scripts/MaxYari/ReAnimation_v2/ReAnimation_v2.lua
+PLAYER NPC: scripts/MaxYari/ReAnimation_v2/Data.lua
